+++ cmd/frontend/internal/graphqlbackend/configuration_cascade.go
@@ -45,10 +45,7 @@ func (r *configurationCascadeResolver) Subjects(ctx context.Context) ([]*configu
 		return mockConfigurationCascadeSubjects()
 	}
 
-	subjects := []*configurationSubject{
-		{site: singletonSiteResolver}, // site config "settings" field
-		{}, // global settings
-	}
+	subjects := []*configurationSubject{{site: singletonSiteResolver}}
 
 	if r.unauthenticatedActor {
 		return subjects, nil
@@ -97,6 +94,18 @@ func (r *configurationCascadeResolver) Merged(ctx context.Context) (*configurati
 		return nil, err
 	}
 	for _, s := range subjects {
+		if s.site != nil {
+			// BACKCOMPAT: Add the site config "settings" field's settings (if any) to the merged
+			// config. They are deprecated but must still be applied.
+			contents, err := s.site.DeprecatedSiteConfigurationSettings()
+			if err != nil {
+				return nil, err
+			}
+			if contents != nil {
+				configs = append(configs, *contents)
+			}
+		}
+
 		settings, err := s.LatestSettings(ctx)
 		if err != nil {
 			return nil, err